Scripts: Adding date-notify script.
I'm experimenting using my window manager without a panel bar, so I
figured I'd like to see the date and time quickly through a
notification. I'm using this with "timeout 5s date-notify.sh" so it
only displays for 5 seconds, and added a rule in my dunst
configuration file (should come to this repo soon) so it times out quickly after that.