emacs: removing embark, company and yasnippet.
Company I've replaced with corfu, and as cool as yasnippet and embark
are, Embark I never really got into the habit of using, even though I
gave it pretty nifty keybindings, and yasnippet I'll have to
reconfigure for corfu, or replace it with some other package.
For now I'm okay without it though.