emacs: Using undo-tree instead of undo-fu
I had switched to undo-fu when I started having a lot of issues with
undo-tree, and undo-fu-session made me stay with it because having an
undo history I could come back to was huge!
But then, undo-fu started letting me down. It's linearity really
started getting in the way, and the "unconstrained" undo/redo just
seemed to never do what I'd expect, and so I went back to see how
undo-tree was doing, and it seems to be doing a lot better! With
undo-tree-history I can keep the behavior from undo-fu-session, and
I'm back to the lovely trees of undo that I really liked, no issues so far.